виніс решту компонентів в окремі файли
